+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 2016-10-12
+* Added 2 environment variables to booster fleet
+
+### 2016-10-10
+* Update aqua version to 1.2.3
+* Move away from behance launcher container
+
 ### 2016-10-10
 * Allow for the configuration of docker logging driver for application deployments.